fix phone number problem

+++ CPP-00/ex01/Contact.hpp
@@ -6,7 +6,7 @@
 /*   By: ademurge <ademurge@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/16 17:36:07 by ademurge          #+#    #+#             */
-/*   Updated: 2022/12/18 20:43:03 by ademurge         ###   ########.fr       */
+/*   Updated: 2023/02/15 11:50:06 by ademurge         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,24 +22,24 @@ public:
 	Contact(void);
 	~Contact(void);
 
-	std::string	getDarkSecret(void) const;
-	std::string	getFirstName(void) const;
-	std::string	getLastName(void) const;
-	std::string	getNickName(void) const;
-	long		getPhoneNumber(void) const;
+	std::string				getDarkSecret(void) const;
+	std::string				getFirstName(void) const;
+	std::string				getLastName(void) const;
+	std::string				getNickName(void) const;
+	unsigned long long		getPhoneNumber(void) const;
 
 	void		setDarkSecret(std::string s);
 	void		setFirstName(std::string s);
 	void		setLastName(std::string s);
 	void		setNickName(std::string s);
-	void		setPhoneNumber(int nb);
+	void		setPhoneNumber(unsigned long long nb);
 
 private:
-	std::string	darkSecret;
-	std::string	firstName;
-	std::string	lastName;
-	std::string	nickName;
-	int			phoneNumber;
+	std::string					darkSecret;
+	std::string					firstName;
+	std::string					lastName;
+	std::string					nickName;
+	unsigned long long			phoneNumber;
 };
 
 #endif
